Make snapshot executable and rename snapshot files
