+++ micro_manager/__init__.py
@@ -2,8 +2,8 @@ import argparse
 import os
 
 from .config import Config
-from .micro_manager import MicroManager
-from .snapshot.snapshot_computation import SnapshotComputation
+from .micro_manager import MicroManagerCoupling
+from .snapshot.snapshot import MicroManagerSnapshot
 
 
 def main():
@@ -22,9 +22,9 @@ def main():
         config_file_path = os.getcwd() + "/" + config_file_path
 
     if not args.snapshot:
-        manager = MicroManager(config_file_path)
+        manager = MicroManagerCoupling(config_file_path)
     else:
-        manager = SnapshotComputation(config_file_path)
+        manager = MicroManagerSnapshot(config_file_path)
 
     manager.initialize()
