+++ micro_manager/snapshot/dataset.py
